RFC: clang-tidy: Enable bugprone-integer-division
Should this be enabled?
Pro: It seems to catch real errors. Makes code more readable.
     (I'm not sure I have fixed all the correct way which means
      the code was hard to read before)
Cons: It has some false postives

If so how do we want to mark attentional int divisions?
(int)(a/b)            #in this patch
static_cast<int>(a/b) #style guide but kind of long for this purpose
int_div(a,b)          #new function for this purpose

+++ src/gromacs/domdec/distribute.cpp
@@ -318,7 +318,7 @@ static void distribute_cg(FILE *fplog,
     snew(tmp_ind, dd->nnodes);
     for (i = 0; i < dd->nnodes; i++)
     {
-        tmp_nalloc[i] = over_alloc_large(cgs->nr/dd->nnodes+1);
+        tmp_nalloc[i] = over_alloc_large((int)(cgs->nr/dd->nnodes)+1);
         snew(tmp_ind[i], tmp_nalloc[i]);
     }
